requirements.txt
pandas
redis
python-dotenv
